Git Repositories View: Refactoring first part
The Git Repositories View should be refactored to use
the Object Contribution extension point for building
the context menus instead of doing this programmatically
in the View class itself.
This is the first part of this refractoring. It does not
refactor the menu contribution yet, just makes makes the previous
tree node class abstract and creates a new concrete subclass
for each tree node type in order to being able to use those
classes in the extension.