Silence a javadoc diagnstic with Eclipse 3.4
The FilteredCheckboxTree.getRefreshJobDelay method overrides
a method in Eclipse 3.5, which means it does not require
a javadoc. The project settings reflects and enforces this,
which does not work with Eclipse 3.4 since getRefreshJobDelay
is a new method there and the compiler flags this as an error.

Since this is the only place where this is a problem we create
a javadoc comment instead of relaxing the rules.

Change-Id: I9861336bb7b81d398119d4355e0fe3d303c38b55
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>