Extract public class RepositoryCache from GitProjectData
GitProjectData contains a cache for Repository instances.
This cache is used for Reposiotry instances related to
projects. Since there are also Repository instances with no
related project (e.g. in RepositoriesView) the cache was moved
to a public class RepositoryCache to enable sharing Repository
instances even if there is no related project.

Change-Id: I5526362279d87fd6f2651ee991bb08ea4cd65c2b
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>