Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve; IP review completed

The version of wikitext that we were using previously was old so we need to make sure we pickup the latest one from Helios... it has the flexibility to grab entries from different wiki pages. I documented this process here:

http://wiki.eclipse.org/DocumentationGuidelines/CrowdSourcingExample

We can greatly simplify the script with the new wikitext to. We have a bug to do this already :)

306205: Update to the latest wikitext
https://bugs.eclipse.org/bugs/show_bug.cgi?id=306205

Which version of wikitext are you fetching with this change?

Patch-set: 1