Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

We need 1.3+ in order to be able to make use of the nice improvements which simplified our ant script considerably. This version is available here
http://download.eclipse.org/tools/mylyn/update/weekly/e3.4

You are right we don't have Java in here, so I will remove the  Java nature.

Patch-set: 1