Git Import Projects: always show directories
Currently, the import projects wizard only shows
the working directory tree if it is started from the
clone wizard. The working directory tree should
also be shown when the wizard is started from the
Git Repositories view in order to improve usability.
If a subfolder of the working directory is selected
for starting this wizard, the selected folder should
be marked in the wizard's tree. The user can still
change the selected folder before hitting "Next".

Bug: 315169

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>