Use informational icon for 'hidden repositories'
When one or more git repositories have been added in the Git Repositories view,
a warning message is shown in the dialog area with 'n directories are hidden as
they have already been added'.

This shouldn't be a warning message with a warning icon; it should be
informational, at best.

Bug: 315558
Change-Id: Ic75cbaf5a88f58c4c01ea68f159f842362df5cd5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>