Use informational icon for 'hidden repositories'
When one or more git repositories have been added in the Git Repositories view,
a warning message is shown in the dialog area with 'n directories are hidden as
they have already been added'.

This shouldn't be a warning message with a warning icon; it should be
informational, at best.

Bug: 315558
Change-Id: Ia0fe21228c52fad72977fabeb2bb38960dd06eb8
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>