Allow Git repository to be pre-filled from clip
If the user has copied a clipboard already, then we can access that
from the SWT clipboard. If it looks like a transport we know about
(as reported by Transport.canHandleProtocol) then we pre-fill the
connection dialog with that as the URI.

Since we need to fill in the user/port combinations, refactor out
the code which does this on keypress and instead call that function,
but only once, and only after all the fields have been created.

Bug 315589