Remember to dispose the clipboard after use
Once the clipboard is acquired, we should dispose of it afterwards to
prevent memory and resource leaks.

Not all the text on the clipboard will be text; if it isn't, then we'll
get a null back. Handle that case specifically.

If the text begins (or ends) with whitespace, trim it before checking.

Bug: 315589
Change-Id: I1b8eeae47b880b6eef689ca359f9c67e93e7e999
Signed-off-by: Alex Blewitt <Alex.Blewitt@gmail.com>