Update patch set 1
Patch Set 1: No score; No score

I think it would be useful if the search button would be the default button when the dialog comes up. This would allow to hit the search button using the return key. As soon as a valid search result is select this should switch to the OK button. The other option would be to have the OK button as default and provide keybord shortcuts for the other buttons.

Also I feel that the distribution of buttons across the dialog could be improved. Currently there are buttons in all corners of the dialog which requires quite a lot of mouse movements to select the diverse buttons. Just see how much you must move the mouse to take the standard path :
- "Browse" (upper right corner) to select directory
- "Search" (upper left corner) to search repositories
- "Select all" (lower left corner) or click checkboxes to select repositories to import
- "OK" or "Cancel" (lower right corner)

Patch-set: 1