Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

Mostly ok, but the commit message should be clearer. It contains a bit too much text actually. Also meantion that
this is not a complete solution since it does not open projects that appear. This would happen due to refactoring. The actual implementation of that can be subject of another patch. In those cases the workspace would still be inconsistent, even when we have all projects.

I also think the user should be warned that closing/opening is about to happen and have the option to abort before anything happens.

We need more Unit tests. Yep, it's hard. Whether you do it using SWTBot or not is less important to me, but it's probably a bit better if you can do without.

Patch-set: 1