Refactor "previously used values" content assist
Currently, there are two places where we have implemented
a content assist with previously used values: in the commit dialog
for the author and commiter fields, and in the RepositorySelection
page for the URI field. The logic should be put into one place and
be documented properly to ensure consistency, supportability,
and re-usability.

Bug: 315186

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>