Update patch set 1
Patch Set 1:

Please remove that blank line between Bug and Signed-off-by.  Its preventing Gerrit from seeing the Bug: 316761 line as a footer, which means it won't index this change as a fix for 316761... and we can't later search Gerrit for changes related to bug 316761 with the search string "bug:316761".  :-)

Patch-set: 1