Update patch set 14
Patch Set 14:

Yes now it is time for next iteration (and fixing unit test for this one ;)). Before it will be merged into master it should be merge it with latest Mathias Kinzler Repositories view so that we can launch synchronize for Repositories view context menu.

I'll do that today evening and push it once again.

Patch-set: 14