Update patch set 3
Patch Set 3: No score; No score

One question to ask is when do we want to do the initial CQ for this? I think it's easier to get what we have when it reaches an acceptable level and then just iterate in <250 line patches to avoid the CQ switch.

What do people think?

Patch-set: 3