Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

IndexOutOfBoundException when I try to delete an entry in the branch column. Afterwards the dialog is not functionable anymore.

The precondition for enabling "Finish" that the same number of branches must be selected as repositories is not very usable. There could be the default "HEAD" for the selected repos.

Patch-set: 1