Update patch set 9
Patch Set 9:

> I would just fix the problem with the platform dependent path separator. In other places e.g. RepositoriesView we also show the full path to the GITFOLDER.

Full patch to the GITFOLDER can make synchornize dialog extremely huge in some cases.

> When playing around I hit the following error quite frequently:

I'll inspect that.

> I couldn't really find out when the synchronize action is activated in the repositories view context menu. Sometimes it's there sometimes its greyed out.

I've already fix it. Right now it is depend on project name (it should be equal with repository name), I've reimplemented that and check does we have any project in workspace that is connected with selected repository

> Could you cover all possible permutations in GitSyncInfo.calculateKind() with tests ?

OK

> I think the GitResourceVariantComparator would also deserve a test.

OK

> In general I think you should think about a way to move most of the logic down into org.eclipse.egit.core and then write core tests for the basic logic. They are faster, and don't depend on brittle label text search logic.

Yes, this is a good idea. I'll try to move as much I can to egit.core.

Patch-set: 9