Update patch set 1
Patch Set 1:

I'm agree with Chris, we should get 'Synchronization View' support as soon as it is possible. I'll prefer move on with Remy's changes when I finish working on tagging support. IMHO there is no need to rewrite this right now especially when we want to have this feature in 0.8.0.

Patch-set: 1