Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(4 inline comments)

The commit message should not explain what you are doing but why you are doing this change. The what is in the diff ...

So I would write something like:

Fix NPE when deleting project used in Synchronize View

NPE is hit when user deletes project that currently is used in Sychronize View.

Patch-set: 1