When creating branch, set current branch as suggested source branch
This is the second part of the solution for bug 315167. If the currently
checked-out branch is a remote branch, the combo will be pre-selected
with this branch when creating a new branch from the "Branches" node
of the Git Repositories view.
As additional convenience, if a remote brach is the suggested source,
the last part of that branch name will be suggested as name for the
local branch.

Bug: 315167
Change-Id: Ica554cea417c2e0d5de86c1cedfabeac737048af
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>