Stop using TeamAction
We should not use the (internal) class TeamAction in order to
implement our actions. We still have to contribute Actions as
Object contributions in the action framework, but we should
put the implemenations of the actions into action handlers
which can then be re-used, for example between the team context
menu actions and the command contributions used in the 
Git Repository view.
This is the first step which splits the current actions contributed
to the team menu into a lightweight action part and a handler part
which is doing the actual work. The next step would be to weed out
the double implementations in the Git Repositories view and either use
the handler defined here or use a second set of handlers and make use
of the "enabledWhen" feature of the "handler" extension point.