Show truncated commit ids when in comparison view
When comparing two resources, the full commit identifier is shown.
This is generally too long (and unnecessary) so truncate this as we
are doing elsewhere.

Make a helper method to return that truncated form so that we can
change the parameterisation if necessary (and/or a preference).
Leave it private in this class until otherwise needed.

Bug: 317376
Change-Id: I87013183359fc1985a20cb5faf95354090c14ac3
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>