Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; IP review completed

Thanks for this fix which looks good. However, I would like you to change the commit message according to the description found in

http://wiki.eclipse.org/EGit/Contributor_Guide#Commit_message_guidelines

Please paste the Change-Id from this change into it so that Gerrit can re-use this change.

Patch-set: 1