Exploit command framework in EGit
Bug: 313589
Bug: 306459
Bug: 314096

We should not use the (internal anyway) class TeamAction in order to
implement our actions. We still have to contribute Actions as
Object contributions in the action framework so that they
appear in the Team context menu, but we should
put the implementing code into command handlers
which can then be re-used, for example between the team context
menu actions and the command contributions used in the 
Git Repository view. The actions are just lightweight objects that
delegate their run() method to the command.

In addition, this is a prerequisite for adding key bindings (it
is possible to add a command id to the contributed actions using the 
"definition id" feature of the corresponding extension point; the
key binding (if found) defined for that command will then automaticall
be displayed in the context menu).

This only the first step which splits the current actions contributed
to the team menu into a lightweight action part and a command handler part
which is doing the actual work. The next step would be to weed out
the double implementations in the Git Repositories view and either use
the handler defined here or use a second set of handlers and make use
of the "enabledWhen" feature of the "handler" extension point to enable
the correct handler depending on the workbench selection.

In a future step, key bindings could be added to the commands.

Change-Id: I910f234e8880bc2e2fbfbd495ac97b63967a38b9
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>