Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Took me a while to figure out what the tt stood for. Would tip be better?

Ideally I'd like this not to be hardcoded but to call a method to get this list on demand - so we can delegate to jgit in the future to find out supporteded protocols. 

Since we are touching this anyway, can we extract the list into a method called "getSupportedMethods" instead of assuming a built-in array? It could return a (tree)map of protocol names (as key) and tooltip as the value.

Patch-set: 1