Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

This looks good to me. I suggest this gets added (and the dependent one abandoned) as a means of fixing the ssh/git+ssh issue simply - then work on a new way of exposing the set of exported protocols from the JGit layer instead of the EGit layer.

Patch-set: 3