Avoid remote connections while validating wizard pages
The bug describes a specific problem which was occurring
in different wizards due to their implemenation using
some event mechanism to notify the "next" wizard page
of URI changes. This has been removed and the "next"
wizard page is notified about the URI by overriding the
setVisible() method and setting the current URI just
before the page opens.
This pattern has been applied for all affected pages.
It also helps in decoupling the different wizard
pages by removing hard-coded usages between them.
As for the specific bug, we may need some other
improvements depending on whether URI validation on
the RepositorySelectionPage is needed. This could for
example be added as a 'Validate URI' button or such.

Bug: 315575
Change-Id: Id001d83db1c042f0016a15f2f3c18c1abcce9dc0
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>