Update patch set 3
Patch Set 3:

My goal was to cleanup this class a bit since protocol intrinsic information was spread over the whole class RepositorySelectionPage. 

I agree that in another step we should further refine this and find a solution to make this part of the JGit API. I also agree that for a public API the currently chosen enum approach might not be optimal. Hence I switched visibility of the enum to private to make clear that this implementation is not part of any public API.

But still I believe this proposal is cleaner than the previous implementation.

Patch-set: 3