Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Oh good grief, not an enum. Enums represent the ultimate evil for potentially extensible data sets as it's impossible to add new irks after the fact. In any case, this should really be a front to a call in JGit rather than permanent code in EGit

Patch-set: 1