Stabilize Core Tests
There are currently failures when the test workspace is not cleared.
In the TrackUntrackOperation test, apparently, the check whether a file is "tracked" was too
simplistic.
Other fixes were required because resources were not handled properly.
With these fixes, it is possible to run the tests without clearing the
test workspace.

Change-Id: I31cbbb63b5aa345d3f4afdea1078640973bbbbac
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>