Generate and run repository used by UI tests inside test harness
The sample repository for testing clone integration tests is now
generated, has a much smaller size and its daemon is run as part
of the test harness to speedup the UI tests and make them
independent from external resources.

Also increase memory settings when running tests from maven
and increase timeout to wait for list of branches in clone wizard.

Change-Id: I9a0c59dfaa7b02cbada43a89736826d14529f565
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>