EGit UI Tests: fix timing problem in Clone tests
With the advent of http://egit.eclipse.org/r/#change,735,
one of the GitCloneWizard tests fails if a "replay delay" is used,
since the "fix" branch of the test repository alone is cloning too fast for the
SWT Bot to capture the window that is displayed uring clone.
In order to make this work again, the "master" branch is also cloned during the
specific test.
Without delay, the test is running fine.

Change-Id: Id9d29a625b6163dab3bd1a660f811f8610346773
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>