EGit UI Tests: fix timing problems
With the advent of http://egit.eclipse.org/r/#change,735,
one of the GitCloneWizard tests fails if a "replay delay" is used,
since the "fix" branch of the test repository alone is cloning too fast for the
SWT Bot to capture the window that is displayed during clone.
On the conrary, without a delay, one of the Git Repositories View
tests fails when checking for the label decorations after check-out.

With these fixes, both the delayed and the undelayed launch
configurations should work for running AllTests (the memory
settings for the launch configurations were also adjusted in order
to accommodate the Synchronize tests).

Change-Id: Id9d29a625b6163dab3bd1a660f811f8610346773
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>