Update patch set 2
Patch Set 2:

HandlerUtil's "checked" methods throws an ExecutionException though that is "valid" because an IHandler's execute(ExecutionEvent) method does declare that it can throw an ExecutionException so you can propagate the exception upwards if you wish.

So feel free to use the "checked" method if a 'null' return value will cause problems.

Patch-set: 2