Update patch set 1
Patch Set 1:

The chance of Team doing anything for 3.7 is low... for now... and if they did anything... it was force us to depend on 3.7 as a minimum requirement (which we don't want to do at the moment). So for 0.9, I think this is the right approach for now.

Patch-set: 1