Add "Show History" and "Show in Repositories" to Team menu
Bug: 315459
Bug: 316274

The generic Show In > History context menu is not shown in all perspectives and the
perspective extension point used to contribute to is must be implemented
for each and every perspective declaratively, so depending on the perspective
the menu items contributed by us will not be available.
Offering Team >Show History is also consistent with CVS and SVN.
We should consider to change Compare With > History in such a way as to
display a "History dialog" (with content similar to the history view) to select a
version which is probably closer to what an end user would expect (at the moment
this is rather the same as Show In > History / Team > Show History).


Change-Id: I6ef64e26810299bb2dcfaa6db83fa46c9dcc2b24
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>