Use project name from project description file
When we have a .project file, we should always create the project
metainformation based on this file. I don't see any reason to use
the file path to determinate the project name. In addition,
Platform.getLocation has nothing to do with the workspace but rather
the working directory of the platform. Also refactored the ProjectRecord
a little bit to simplify it. This also fixes bug 314035 to use the project
name of the .project file.


Bug: 314035
Change-Id: I589c510550946be9b92d00746c3f4bb91447d515
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>