Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Could you rebase this? It conflicts with the Export-Package of http://egit.eclipse.org/r/#change,1238 and doesn't compile as you deleted a package.

And I wonder if we should add a checkbox to the "Create branch" dialog to checkout this branch immediatly. Creating a branch in the Git repo view means branch+checkout while in the History view, we only do a branch. Confused me a little bit when I used it the first time. What do you think?

And I get a NPE when using Checkout on a commit in one of my sample project:
java.lang.NullPointerException
	at org.eclipse.egit.ui.internal.actions.CheckoutCommitActionHandler.execute(CheckoutCommitActionHandler.java:140)
	at org.eclipse.egit.ui.internal.actions.RepositoryAction.run(RepositoryAction.java:98)
	at org.eclipse.ui.internal.PluginAction.runWithEvent(PluginAction.java:251)

Patch-set: 2