Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Hm, can't seem to reproduce the NPE at will. Currently I'm unable to find a scenario to reproduce it. Maybe there was something wrong with my test project. I think we should leave it as is (tested it with several scenarios) and if it happens again, I'll open a bug.

Another change for the checkbox: +1

Patch-set: 4