Update patch set 5
Patch Set 5:

Dariusz, perhaps you can 'bisect' the failing tests?

Start with running the failing test with all the tests, and then split the test suite into two. Run the GitResourceVariantTreeTest with one half. If it fails, you have eliminated half the tests as part of the failing combination. If it does not, run the failing test with the other half of the tests.

Wash, rinse, repeat and you'll know which minimal combination of tests fails, which should bring you closer to discovering the root issue.

Patch-set: 5