Update patch set 8
Patch Set 8: I would prefer that you didn't submit this

(2 inline comments)

I think that in GitSyncInfo we do not need anything (E)Git specific. I would even assume that we could use plain SyncInfo. The handling of folders could be done in the Comparator (see ThreeWayResourceComparator as an example).

If that is done or it is not feasible to do it I would say we could finally submit the change.

Patch-set: 8