Update patch set 4
Patch Set 4: (6 inline comments)

Indeed unclosed streams was an issues in GitResourceVariantTreeTest, but they didn't cause failures in GitResourceVariantTreeSubscriberTest. This problem still exists.

I know about issue with synchronization when source branch isn't same as actually check-outed. But (as I write in comment) I'll prefer to fix it (or make a workaround for it) as a separate commit. *Maybe* when we will be integrated with Logical Model this issue can be solved properly.

Patch-set: 4