Git History View: cleanup and usability
This change is intended as first step to imporove the usabily of the
Git History View. From the end user perspective, the following changes
are noticeable:

1. there is a new toolbar button "Show all versions for the resource"
which makes the "resource mode" explicit (in other words, we now have
four radio buttons insted of three where selection of no button at all
would represent the fourth option)
2. "Wrap comments" and "Fill paragraphs" is only available when the context menu
is opened on the comment viewer
3. The file-specific menu entries (compare, open) were promoted to the top of the context
menu if a file is selected

Internally, all the "popupMenu" contributions were removed and replaced
with command contributions and the corresponding actions were refactored out
of the GitHistoryView class into handlers in order to improve maintainability.

Change-Id: I318476873c0142e18e3cf8928ea693f3adaba1ab
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>