Update patch set 4
Patch Set 4: I would prefer that you didn't submit this; No score

The ignore logic should be mostly in JGit, with
hooks that EGit can use.

The current ignore handling was a quick-fix because it
was very easy to implement and .gitignore is hard to implement.

There is a reason we don't have .gitignore yet. Doing it properly is hard and we don't want it if it doesn't 
work exactly as the C Git version, hence we don't support it at all (for the time being), which is better than having an incompatible implementation.

The complexity means we need proper unit tests for it. I do like the structured approach for definiing the test cases, which means implementing unit tests should be rather straightforward.

There have beens some attempts that were rejected. They are pre-git and the mails can be found in the GITM archives. Search for JGit / EGit and ignore

Patch-set: 4