Trigger repository refresh independently of workspace refresh setting
Initially we always refreshed when the repository changed, but users
we not happy, so we tied it to the workspace refresh option. This means
Eclipse performs unnecessary work by periodically scanning all recources
as well as scanning when Git refs or index change.

This patch allows Eclipse to rescan only when the refs or index change.
Eclipse still scans some of the repository meta data, but that is a lot
fewer files.

In addition the default is to only scan for repositirt changes or refresh
resources when Eclipse has an active Windows, which means that you can
perform Git operations outside of Eclipse with no (or at least less)
interaction between Eclipse and C Git. This used to be a big problem
for Windows users performing rebase operations.

Change-Id: Id3fef6b884509e1a1a0bbf62fd1ae40ee0f1211e
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>