Trigger repository refresh independently of workspace refresh setting
Initially we always refreshed when the repository changed, but users
we not happy, so we tied it to the workspace refresh option. This ment
Eclipse performed unnecessary work by periodically scanning all recources
as well as scanning when Git refs or index changed.

This patch allows Eclipse to rescan only when the refs or index change.
Eclipse still scans some of the repository meta data, but that is a lot
fewer files.

In addition the default is to only scan for repository changes or refresh
resources when Eclipse has an active window, which means that you can
perform Git operations outside of Eclipse with no (or at least less)
interaction between Eclipse and C Git. This used to be a big problem
for Windows users performing rebase operations, because Eclipse locked
files for a brief period prevening C Git from updating them.

Change-Id: Id3fef6b884509e1a1a0bbf62fd1ae40ee0f1211e
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>