History View: improve "enablement" calculation.
With the new command-based menu contributions implemented for
the history view, we get lots of

"Could not get the current input from the history view"

messages in the log. This is because the enablement if calculated
even if the history is not visible at all.

This change does the calculation properly.

Change-Id: I113532b98350434034564f03f9ab05627c4663e6
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>