Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

(9 inline comments)

Please habe a look at the inline comments.

The labels of the left and right compare editor inputs after double clicking a member of a change set are missing. They would be very useful. It took me quite a time to find out how to get them in, so I'll just push this change to gerrit.

Patch-set: 5