Use DiffFormatter inside FileDiff
Make use of the jgit classes DiffEntry and DiffFormatter to calculate 
diff properties from a tree walk and to format the diff.

Unfortunately the DiffFormatter cannot yet be used for the eclipse
patch. The eclipse patch format does not show any source or 
destination prefix,  and the paths are calculated relative to the eclipse
project.

Bug: 320966
Change-Id: I18bc942208f64abbcf8adb482f84fcb305194782
Signed-off-by: Stefan Lay <stefan.lay@sap.com>