Trigger repository refresh independently of workspace refresh setting
Initially we always refreshed when the repository changed, but users
we not happy, so we tied it to the workspace refresh option. This ment
Eclipse performed unnecessary by scanning resources twice. Both
an infrequent periodical scan and a triggered scan.

Now we only scane for index changes since a ref change by itself does
not hint that a resource has changed. In addition the default is to
only scan for repository changes and refresh resources when Eclipse
has an active workbench window, which means that you can perform Git
operations outside of Eclipse with no (or at least less) interaction
between Eclipse and C Git. This used to be a big problem for Windows
users performing rebase operations, because Eclipse locked files for
brief periods preventing C Git from updating them.

Only externally initiated changes are cared for since we are only looking
for a cheap way of detecting resource changes when git operations
are performed outside of Eclipse, mainly rebase and merge.

Change-Id: Id3fef6b884509e1a1a0bbf62fd1ae40ee0f1211e
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>