Workaround for bug 322751
Since we can't fix the problems described in bug 322751,
this suggests are workaround to set the workbench selection
in a reliable way.
Ugly enough, it forces a hide/activate cycle on the
history view if it is not currently holding a GitHistoryPage.
This will enable the selection provider correctly.
This can only be done on the Team->Show History
context menu item, so we will have to remove the
generic Show In->History action where this workaround
is not available.

Bug: 322751
Change-Id: I1375afc763f15b0e5a1b932bbb1ae24694ea6428
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>