Update patch set 2
Patch Set 2:

Chris, Paul: thanks a lot for following up here. As far as I could see, the issue is that the generic history view is activated and only then the Git History page is initialized which in turn is trying to set the selection provider (on the part site). This is too late, as the selection service is registering selection change listeners to the selection provider while activating the generic history view).
Anyway, this change does away with this anyway.

Patch-set: 2