Update patch set 1
Patch Set 1:

Talking to Paul Webster earlier... he mentioned this...

"It seems like the getSelection(*) in your handler should be updated. Have it get the activeMenuSelection first, and only try "selection" if activeMenuSelection is not available.

You still need "selection" if the command has to respond to keybindings, since activeMenuSelection is only available while the context menu itself is shown."

Patch-set: 1