Git Configuration UI
The "Properties" view shows the configuration of the repository
currently selected in the repositories view. The toolbar was cleaned up
and simplified and the "direct editing" available in the properties view
was removed. Instead, when hitting a new "Edit" button in the toolbar, a 
dialog appears presenting the same UI that is used in the "Global Git Configuration"
preference page which can then be used to edit the configuration in a comfortable
way.

Change-Id: I9c15f03379683fc61c61fcf53cc2668883f12bc3
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>