Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve; IP review completed

This is sooo cool! I'm just not sure if the usage of the JGit API is correct, so someone else should have a look. Also I think we need to find a better icon, but that's something we should discuss with Eitan (our UI guy at SAP) anyway in the course of the greater usability discussion around the history view.

Patch-set: 2